requirements.txt
Flask==2.2.5
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==20.1.0
matplotlib
seaborn
xgboost
lightgbm
dill
langchain
langchain_groq
python-dotenv
faiss-cpu
langchain_community
duckduckgo-search
Wikipedia
-e .